+++ packages/extension-code-block-lowlight/src/lowlight-plugin.ts
@@ -37,6 +37,12 @@ function getDecorations({ doc, name, lowlight }: { doc: ProsemirrorNode, name: s
         ? lowlight.highlight(language, block.node.textContent).children
         : lowlight.highlightAuto(block.node.textContent).children
 
+      console.log({
+        lowlight,
+        languages,
+        nodes,
+      })
+
       parseNodes(nodes).forEach(node => {
         const to = from + node.text.length